requirements_windows.txt
asttokens==2.4.1
comm==0.2.2
debugpy==1.8.9
decord==0.6.0
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
importlib-metadata==8.5.0
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.3.6
prompt-toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pygments==2.18.0
python-dateutil==2.9.0.post0
pyzmq==26.2.0
six==1.16.0
stack_data==0.6.2
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
wheel==0.45.0
zipp==3.21.0
absl-py==2.1.0
accelerate==1.1.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.2.0
av==13.1.0
black==23.7.0
blinker==1.9.0
braceexpand==0.1.7
cachetools==5.5.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.7
git+https://github.com/openai/CLIP.git
cmake==3.25.0
contourpy==1.3.1
cycler==0.12.1

diffusers==0.31.0
docker-pycreds==0.4.0
einops==0.8.0
fairscale==0.4.13
fastapi==0.115.5
ffmpeg-python==0.2.0
ffmpy==0.4.0
filelock==3.13.1
fire==0.7.0
flatbuffers==24.3.25
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
ftfy==6.3.1
future==1.0.0
gast==0.6.0
gcsfs==2024.10.0
gitdb==4.0.11
gitpython==3.1.43
google-api-core==2.23.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gradio==5.6.0
gradio-client==1.4.3
grpcio==1.68.0
h11==0.14.0
h5py==3.12.1
hjson==3.1.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.4
imageio==2.36.0
invisible-watermark==0.2.0
jinja2==3.1.3
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.6.0
kiwisolver==1.4.7
kornia==0.6.9
lazy-loader==0.4
libclang==18.1.1
lightning-utilities==0.11.9
lit==15.0.7
llvmlite==0.43.0
markdown==3.7
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.14.1
natsort==8.4.0
networkx==3.2.1
ninja==1.11.1.1
numba==0.60.0
numpy==1.26.3
oauthlib==3.2.2
omegaconf==2.3.0
open-clip-torch==2.29.0
opencv-python==4.6.0.66
opencv-python-headless==4.10.0.84
opt-einsum==3.4.0
optree==0.13.1
orjson==3.10.11
pandas==2.2.3
pathspec==0.12.1
pillow==10.2.0
piqa==1.3.2
pooch==1.8.2
propcache==0.2.0
proto-plus==1.25.0
protobuf==3.20.3
pudb==2024.1.3
py-cpuinfo==9.0.0
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.1
pydantic-core==2.27.1
pydeck==0.9.1
pydeprecate==0.3.1
pydub==0.25.1
pymatting==1.1.13
pynvml==11.5.3
pyparsing==3.2.0
python-multipart==0.0.12
pytorch-lightning==1.5.0
# pytorch-kmedoids is unused and lacks Python 3.12 wheels
# pytorch-kmedoids==0.0.1
pytz==2024.2
pywavelets==1.7.0
pyyaml==6.0.2
referencing==0.35.1
regex==2024.11.6
rembg==2.0.60
requests==2.28.1
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
ruff==0.8.0
safehttpx==0.1.1
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1

semantic-version==2.10.0
sentry-sdk==2.19.0
setproctitle==1.3.4
shellingham==1.5.4
smmap==5.0.1
sniffio==1.3.1
starlette==0.41.3
streamlit==1.40.1
streamlit-keyup==0.2.0
sympy>=1.13.3
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardx==2.6
tensorflow==2.18.0
# tensorflow-io-gcs-filesystem has no Python 3.12 wheels; not needed unless using GCS
# tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
test-tube==0.7.5
tifffile==2024.9.20
timm==1.0.11
tokenizers>=0.21,<0.22
toml==0.10.2
tomli==2.1.0
tomlkit==0.12.0
torchmetrics>=1.2.0
tornado==6.4.2
tqdm==4.67.0
transformers==4.47.0
typer==0.13.1
tzdata==2024.2
urllib3==1.26.13
urwid==2.6.16
urwid-readline==0.15.1
uvicorn==0.32.1
wandb==0.18.7
watchdog==6.0.0
webdataset==0.2.100
websockets==12.0
werkzeug==3.1.3
wrapt==1.17.0
# xformers is installed separately by install_windows.bat to ensure cu128 compatibility
# xformers==0.0.33.post2
yarl==1.18.0
